Добавил удаление типа переплета
