+++ react-web-client/src/components/admin-panel/cover-type/CoverTypeForm.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import cs from './../../body/content/Content.module.css';
 import AddCoverType from './add-cover-type/AddCoverType';
 import UpdateCoverType from './update-cover-type/UpdateCoverType';
+import DeleteCoverType from './delete-cover-type/DeleteCoverType';
 
 
 const CoverTypeForm = (props) => {
@@ -10,7 +11,7 @@ const CoverTypeForm = (props) => {
 
         <div className={cs.content}>
             <AddCoverType addCoverType={props}/>
-            {/* <DeletePainter deletePainter={props}/> */}
+            <DeleteCoverType deleteCoverType={props}/>
             <UpdateCoverType updateCoverType={props}/>
         </div>
     );
