+++ react-web-client/src/components/admin-panel/cover-type/delete-cover-type/DeleteCoverType.css
