+++ react-web-client/src/render.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
-import { setPainterStyle, initializeData, deletePainterStyle, updatePainterStyle, setPainter, deletePainter, updatePainter, getPainterByCount, addPublisher, getPublishersByCount, deletePublisher, updatePublisher, getCoverTypes, addCoverType, updateCoverType } from './redux/state';
+import { setPainterStyle, initializeData, deletePainterStyle, updatePainterStyle, setPainter, deletePainter, updatePainter, getPainterByCount, addPublisher, getPublishersByCount, deletePublisher, updatePublisher, getCoverTypes, addCoverType, updateCoverType, deleteCoverType } from './redux/state';
 
 export let rerenderEntireTree = (state) => {
   let painterMethods = {
@@ -23,12 +23,13 @@ export let rerenderEntireTree = (state) => {
   let coverTypeMethods = {
     getCoverTypes,
     addCoverType,
-    updateCoverType
+    updateCoverType,
+    deleteCoverType
   }
 
   ReactDOM.render(
     <React.StrictMode>
-      <App state={state} setPainterStyle={setPainterStyle} deletePainterStyle={deletePainterStyle} updatePainterStyle={updatePainterStyle} painterMethods={painterMethods} publisherMethods={publisherMethods} coverTypeMethods={coverTypeMethods}/>
+      <App state={state} setPainterStyle={setPainterStyle} deletePainterStyle={deletePainterStyle} updatePainterStyle={updatePainterStyle} painterMethods={painterMethods} publisherMethods={publisherMethods} coverTypeMethods={coverTypeMethods} />
     </React.StrictMode>,
     document.getElementById('root')
   );
